+++ src/middleware/permissions.js
@@ -4,34 +4,47 @@ const { auth, messages: errMessages } = require('../constants/errors.json')
 const { rejectIfNull } = require('../db/utils')
 const { oauthServer } = require('../controllers/oauth')
 
-const checkUserIsPermitted =
-  ({ minSufficientRole, resource, resourceOwnerIsSufficient }) => (user, resourceId) => {
-    const orderedRoles = [roles.BASIC, roles.ADMIN, roles.SUPER]
+const hasSufficientRole = ({ minSufficientRole }) => user => {
+  const orderedRoles = [roles.BASIC, roles.ADMIN, roles.SUPER]
 
+  // throw error on initialisation if passed option is not valid
+  if (!orderedRoles.includes(minSufficientRole)) {
+    throw boom.badImplementation()
+  }
+
+  return orderedRoles.indexOf(user.role) >= orderedRoles.indexOf(minSufficientRole)
+}
+
+const checkUserOwnsResource = ({ resourceType, resourceId }) => user => {
+  resourceType.findById(resourceId)
+  .then(rejectIfNull(errMessages.GET_ID_NOT_FOUND))
+  .then(doc => doc.user === user.id
+    ? Promise.resolve()
+    : Promise.reject(boom.unauthorized(auth.UNAUTHORIZED))
+  )
+}
+
+module.exports =
+  ({ minSufficientRole, resourceType, owningResourceIsSufficient }) => (req, res, next) => {
     // throw error on initialisation if passed options are not valid
-    if (!orderedRoles.includes(minSufficientRole)) {
-      throw boom.badImplementation()
-    }
-    if (resourceOwnerIsSufficient && !resource) {
+    if (owningResourceIsSufficient && !resourceType) {
       throw boom.badImplementation()
     }
 
-    if (orderedRoles.indexOf(user.role) >= orderedRoles.indexOf(minSufficientRole)) {
-      return Promise.resolve()
-    } else if (!resourceOwnerIsSufficient) {
-      return Promise.reject(boom.unauthorized(auth.UNAUTHORIZED))
+    if (!req.user) {
+      return next(boom.unauthorized(auth.UNAUTHORIZED))
+    } else if (hasSufficientRole({ minSufficientRole })(req.user)) {
+      return next()
+    } else if (!owningResourceIsSufficient) {
+      return next(boom.unauthorized(auth.UNAUTHORIZED))
     } else {
-      return resource.findById(resourceId)
-      .then(rejectIfNull(errMessages.GET_ID_NOT_FOUND))
-      .then(doc => doc.user === user.id
-        ? Promise.resolve()
-        : Promise.reject(boom.unauthorized(auth.UNAUTHORIZED))
-      )
+      const resourceId = req.params.id
+      return checkUserOwnsResource({ resourceType, resourceId })(req.user)
+      .then(() => next())
+      .catch(err => next(err))
     }
   }
 
-module.exports = opts => (req, res, next) => {
-  checkUserIsPermitted(opts)(req.user, req.params.id)
-  .then(() => next())
-  .catch(err => next(err))
-}
+// export functions for testing
+module.exports.hasSufficientRole = hasSufficientRole
+module.exports.checkUserOwnsResource = checkUserOwnsResource
